Fixed implementation of free() in cli/libs/fileutils.py.
The code as previously implemented did not produce consistent results
across different versions of linux. It worked as probably intended
(there is no documentation in the code as to what it is supposed
to accomplish) in Ubuntu 14.04 but not in Ubuntu 16.04. Switched to
psutil library instead of parsing the output of calls to the shell.